+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+
+## [0.2.0] - 2022-11-13
+### [Added]
 - EEPROM dump support, fix #12
 - Refactor all subcommands, using clap v4
 - Probe support, multiple chips can be selected by an index